flag - & ll

+++ cursus_42/algo/ft_printf/ft_printf.h
@@ -24,7 +24,7 @@ void			ft_manage_flag(char **s, t_flag *flag);
 char			*ft_manage_conv_flag(char c, t_flag *flag);
 char			*ft_manage_str(char c, char *format, va_list *arg, t_flag *flag);
 char			*ft_manage_conv(char c, va_list *arg, t_flag *flag);
-char			*ft_itoa_base_u(unsigned long nb, unsigned long base);
+char			*ft_itoa_base_u(unsigned long nb, int base);
 void			ft_init_flag(t_flag *flag);
 char			*ft_width(char conv, char *format, t_flag *flag);
 char			*ft_join_free(char *s1, char *s2, int op);
